+++ example/android/app/src/main/kotlin/org/tonycloud/flutter_fd_utils_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package org.tonycloud.flutter_fd_utils_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()