+++ example/linux/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,15 @@
+//
+//  Generated file. Do not edit.
+//
+
+// clang-format off
+
+#include "generated_plugin_registrant.h"
+
+#include <flutter_fd_utils/flutter_fd_utils_plugin.h>
+
+void fl_register_plugins(FlPluginRegistry* registry) {
+  g_autoptr(FlPluginRegistrar) flutter_fd_utils_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "FlutterFdUtilsPlugin");
+  flutter_fd_utils_plugin_register_with_registrar(flutter_fd_utils_registrar);
+}